requirements.txt
kafka-python==2.0.2
python-dotenv==0.19.0
argparse==1.4.0